# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.ResponseHTTP:
    properties:
      data: {}
      details:
        type: string
      success:
        type: boolean
    type: object
  common.ResponsePagination:
    properties:
      data: {}
      details:
        type: string
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
...
  models.AppGet:
    description: AppGet type information
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      uuid:
        type: string
    type: object
  models.AppPost:
    description: AppPost type information
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  models.Endpoint:
    description: App type information
    properties:
      description:
        type: string
      feature_id:
        type: number
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      route_path:
        type: string
    type: object
...
  models.Feature:
    description: App type information
    properties:
      active:
        type: boolean
      description:
        type: string
      endpoints:
        items:
          $ref: '#/definitions/models.Endpoint'
        type: array
      id:
        type: integer
      name:
        type: string
      role:
        type: number
    type: object
...
  models.Page:
    description: App type information
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
...
  models.Role:
    description: App type information
    properties:
      active:
        type: boolean
      app:
        type: number
      description:
        type: string
      features:
        items:
          $ref: '#/definitions/models.Feature'
        type: array
      id:
        type: integer
      name:
        type: string
      pages:
        items:
          $ref: '#/definitions/models.Page'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.User:
    description: App type information
    properties:
      date_registered:
        type: string
      disabled:
        type: boolean
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      uuid:
        type: string
    type: object
...
info:
  contact: {}
  description: This is blue-admin API OPENAPI Documentation.
  termsOfService: http://swagger.io/terms/
  title: Swagger blue-admin API
  version: "0.1"
paths:
  /app:
    get:
      consumes:
      - application/json
      description: Get Apps
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponsePagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.AppGet'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      - Refresh: []
      summary: Get Apps
      tags:
      - Apps
    post:
      consumes:
      - application/json
      description: Add App
      parameters:
      - description: Add App
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/models.AppPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.AppPost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add a new App
      tags:
      - Apps
  /app/{app_id}:
    delete:
      consumes:
      - application/json
      description: Remove app by ID
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Remove App by ID
      tags:
      - Apps
    get:
      consumes:
      - application/json
      description: Get app by ID
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.AppGet'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get App by ID
      tags:
      - Apps
    patch:
      consumes:
      - application/json
      description: Patch App
      parameters:
      - description: Patch App
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/models.AppPost'
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.AppPost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Patch App
      tags:
      - Apps
...
  /roleapp/{role_id}:
    delete:
      consumes:
      - application/json
      description: Delete App Role
      parameters:
      - description: App ID
        in: path
        name: role_id
        required: true
        type: integer
      - description: Role ID
        in: query
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Delete App Role
      tags:
      - Apps
    patch:
      consumes:
      - application/json
      description: Add App to Role
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: integer
      - description: ' App ID'
        in: query
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add App to Role
      tags:
      - Apps
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: X-APP-TOKEN
    type: apiKey
  Refresh:
    description: Description for what is this security definition being used
    in: header
    name: X-REFRESH-TOKEN
    type: apiKey
swagger: "2.0"